requirements.txt
Flask==1.1.2
requests==2.24.0
numpy==1.19.0
Werkzeug==1.0.1
torchvision==0.6.0
torch==1.5.0
Pillow==7.2.0
